chore: restructure src/ into lib/, config/, utils/ directories
- Move embedder.ts, walker.ts, languages.ts → src/lib/
- Move env.ts → src/config/
- Move test files alongside their source in src/lib/
- utils/logger.ts already in place
- Update all import paths across source and test files

Closes #18
